+++ src/netguard/include/fd_util.h
@@ -1,0 +1,11 @@
+//
+// Created by Aitor Viana on 25/05/2023.
+//
+
+#ifndef NETGUARD_FD_UTIL_H
+#define NETGUARD_FD_UTIL_H
+
+int is_readable(int fd);
+int is_writable(int fd);
+
+#endif //NETGUARD_FD_UTIL_H
