+++ src/netguard/util.c
@@ -25,7 +25,6 @@
 #include <errno.h>
 #include <netinet/in.h>
 #include <netinet/tcp.h>
-#include <poll.h>
 
 static uint8_t char2nible(const char c);
 
@@ -141,29 +140,6 @@ int32_t get_local_port(const int sock) {
         return -1;
     } else
         return ntohs(sin.sin_port);
-}
-
-int is_event(int fd, short event) {
-    struct pollfd p;
-    p.fd = fd;
-    p.events = event;
-    p.revents = 0;
-    int r = poll(&p, 1, 0);
-    if (r < 0) {
-        log_print(PLATFORM_LOG_PRIORITY_ERROR, "poll readable error %d: %s", errno, strerror(errno));
-        return 0;
-    } else if (r == 0)
-        return 0;
-    else
-        return (p.revents & event);
-}
-
-int is_readable(int fd) {
-    return is_event(fd, POLLIN);
-}
-
-int is_writable(int fd) {
-    return is_event(fd, POLLOUT);
 }
 
 long long get_ms() {
